+++ CHANGELOG.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-Nothing currently!
+### Added
+
+* Added a new `is_terraform_cloud` property to the `Terraform` class (by [@cooperwalbrun](https://github.com/cooperwalbrun))
+
+### Fixed
+
+* Fixed the implementation of `--execution-mode` to properly handle the `agent` mode via inclusion of an optional `agent-pool-id` (by [@cooperwalbrun](https://github.com/cooperwalbrun))
 
 ## v0.5.1 - 2020-12-13